requirements.txt
streamlit
requests
faiss-cpu
sentence-transformers
python-docx
